+++ src/kernel/sample.cpp
@@ -84,8 +84,8 @@ const Sample::TSubSequence1D Sample::subSequence1D(int id) const
 		return TSubSequence1D();
 	}
 	LASS_ASSERT(sampler_);
-	const unsigned k = sampler_->subSequenceOffset1D(id);
-	const unsigned n = sampler_->subSequenceSize1D(id);
+	const size_t k = sampler_->subSequenceOffset1D(id);
+	const size_t n = sampler_->subSequenceSize1D(id);
 	return TSubSequence1D(&subSequences1D_[0] + k, &subSequences1D_[0] + k + n);
 }
 
@@ -98,8 +98,8 @@ const Sample::TSubSequence2D Sample::subSequence2D(int id) const
 		return TSubSequence2D();
 	}
 	LASS_ASSERT(sampler_);
-	const unsigned k = sampler_->subSequenceOffset2D(id);
-	const unsigned n = sampler_->subSequenceSize2D(id);
+	const size_t k = sampler_->subSequenceOffset2D(id);
+	const size_t n = sampler_->subSequenceSize2D(id);
 	return TSubSequence2D(&subSequences2D_[0] + k, &subSequences2D_[0] + k + n);
 }
